Render Question Group
closes #3.

# Create Entities
Models that are not mapped to the database.
- Question
- Answer
- Question Group (or Form)

Different types of question could be created (text, checkbox, radio).
Multiple questions could be added to the Question Group. 
Questions must not be reused by different Question Groups.

# Render the single Question Group
Use __Twig__ library and __Form__ Symfony components.
Render the form with predefined questions list.